[#6] [#7] Do not save NERDTree and Scratch temp files when saving all changed buffes
This fixes #6 and #7

`wa` writes all opened buffers to disk. Even NERDTree and Scratch
buffer, which are not modifiable.

`set autowriteall` saves all buffers if some specific command triggered

Function BufDo found here
http://vim.wikia.com/wiki/Run_a_command_in_multiple_buffers
and lets run command in all buffers, but, unlike standard bufdo, will
be finished in the same buffer when it was started.

So on FocustLost, instead of writing all buffers with `wa` we refresh
all buffers with `e`, which cause their saving.